LayoutTests/accessibility/text-stitching-after-atomic-inline.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN"><!-- webkit-test-runner [ AccessibilityTextStitchingEnabled=true ] -->
<html>
<head>
<meta charset="utf-8"></meta>
<script src="../resources/accessibility-helper.js"></script>
<script src="../resources/js-test.js"></script>
</head>
<body>

<!-- Text after an atomic inline (button) should still be stitched. -->
<div id="test1"><button>Foo</button> Hello <span>team!</span></div>

<!-- Multiple atomic inlines shouldn't prevent stitching between them. -->
<div id="test2">Before <button>A</button> middle <em>text</em> here <button>B</button> after <b>end</b></div>

<script>
var output = "This test ensures text stitching works after atomic inline boxes (like buttons).\n\n";

if (window.accessibilityController) {
    // Test 1: Text after a button should be stitched together.
    var test1 = accessibilityController.accessibleElementById("test1");
    output += "Test 1: text after button\n";
    output += dumpAXSearchTraversal(test1) + "\n";

    // Test 2: Text between and after multiple buttons should stitch separately.
    var test2 = accessibilityController.accessibleElementById("test2");
    output += "Test 2: text between multiple buttons\n";
    output += dumpAXSearchTraversal(test2) + "\n";

    debug(output);
}
</script>
</body>
</html>
